% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.data.r
\name{fetch.data}
\alias{fetch.data}
\title{Download a target power library}
\usage{
fetch.data(alpha, delta, prin = FALSE)
}
\arguments{
\item{alpha}{value of nominal size of test}

\item{delta}{value of clinically relevant difference}

\item{prin}{shows progress of calculation}
}
\value{
No return value. A library object will appear in the global environment.
}
\description{
Function downloads one of 85 power libraries from chrislloyd.com.au.
The libraries are all four column matrices with roughly 20 million
rows and will be around 1.4Gb within R. Download should take roughly
a minute. The object will have a name of the form LIB.alpha.delta.Rdata.
}
\examples{
#'
\dontrun{
fetch.data(alpha=0.05,delta=0.10)
}

}
\references{
C.J. Lloyd & R. Ripamonti (2021) A comprehensive open-source
library for exact required sample size in binary clinical trials.
Contemporary Clinical Trials 107. \doi{10.1016/j.cct.2021.106491}
}
\author{
Chris J. Lloyd
}
